requirements.txt
paho-mqtt
aiohttp
yarl
frozenlist
requests
aiohttp
psutil
pillow